time and hall label on schedule items
