+++ lib/views/client/main/schedule/movie_card.dart
@@ -1,0 +1,291 @@
+import 'package:cached_network_image/cached_network_image.dart';
+import 'package:filmu_nams/assets/theme.dart';
+import 'package:filmu_nams/models/movie.dart';
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:intl/intl.dart' as intl;
+import 'package:loading_animation_widget/loading_animation_widget.dart';
+
+class MovieCard extends StatelessWidget {
+  const MovieCard({
+    super.key,
+    required this.data,
+    this.time,
+    this.hall,
+  });
+
+  final MovieModel data;
+  final DateTime? time;
+  final int? hall;
+
+  @override
+  Widget build(BuildContext context) {
+    final double topMargin = data.title.length > 12 ? 88 : 75;
+    final double bottomMargin = data.title.length > 12 ? 63 : 70;
+
+    return Container(
+      clipBehavior: Clip.antiAlias,
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(10),
+      ),
+      child: Row(
+        textDirection: TextDirection.rtl,
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Stack(
+            clipBehavior: Clip.none,
+            children: [
+              Container(
+                width: 180,
+                margin: EdgeInsets.only(
+                  bottom: bottomMargin,
+                  top: topMargin,
+                ),
+                decoration: BoxDecoration(
+                  color: Theme.of(context)
+                      .bottomNavigationBarTheme
+                      .backgroundColor,
+                  borderRadius: BorderRadius.only(
+                    topRight: Radius.circular(10),
+                    bottomRight: Radius.circular(10),
+                  ),
+                ),
+              ),
+              Positioned(
+                left: 0,
+                top: 10,
+                right: -10,
+                child: Column(
+                  children: [
+                    Padding(
+                      padding: const EdgeInsets.only(bottom: 20),
+                      child: title(),
+                    ),
+                    duration(),
+                    genre(),
+                    director(),
+                  ],
+                ),
+              ),
+              Positioned(
+                bottom: 5,
+                left: -10,
+                right: 0,
+                child: button(),
+              ),
+            ],
+          ),
+          Container(
+            decoration: BoxDecoration(
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.black.withAlpha(150),
+                  blurRadius: 15,
+                  offset: const Offset(5, 0),
+                ),
+              ],
+            ),
+            child: Stack(
+              alignment: Alignment.topRight,
+              children: [
+                poster(),
+                rating(context),
+                if (time != null) scheduledTime(context)
+              ],
+            ),
+          )
+        ],
+      ),
+    );
+  }
+
+  String capitalize(String s) => s[0].toUpperCase() + s.substring(1);
+  String getDuration() => '${data.duration ~/ 60}h ${data.duration % 60}min';
+  String getTime() =>
+      intl.DateFormat(intl.DateFormat.HOUR24_MINUTE).format(time!);
+
+  Container poster() {
+    return Container(
+      clipBehavior: Clip.antiAlias,
+      width: 190,
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.only(
+          topLeft: Radius.circular(10),
+          bottomLeft: Radius.circular(10),
+        ),
+      ),
+      child: CachedNetworkImage(
+        imageUrl: data.posterUrl,
+        fit: BoxFit.cover,
+        placeholder: (context, url) => Center(
+          child: LoadingAnimationWidget.staggeredDotsWave(
+            color: Colors.white,
+            size: 100,
+          ),
+        ),
+      ),
+    );
+  }
+
+  Container rating(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(
+        top: 10,
+      ),
+      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 5),
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.only(
+          topLeft: Radius.circular(10),
+          bottomLeft: Radius.circular(10),
+        ),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withAlpha(100),
+            blurRadius: 10,
+            offset: const Offset(-5, 0),
+          )
+        ],
+        color: red002,
+      ),
+      child: Text(
+        data.rating,
+      ),
+    );
+  }
+
+  scheduledTime(BuildContext context) {
+    return Positioned(
+      bottom: 0,
+      left: 0,
+      child: Container(
+        margin: const EdgeInsets.only(
+          bottom: 15,
+        ),
+        padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 5),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.only(
+            topRight: Radius.circular(10),
+            bottomRight: Radius.circular(10),
+          ),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withAlpha(250),
+              blurRadius: 15,
+              offset: const Offset(5, 5),
+            )
+          ],
+          color: red002,
+        ),
+        child: Text(
+          '${getTime()}  -  $hall. Zāle',
+        ),
+      ),
+    );
+  }
+
+  title() {
+    return TextContainer(
+      Text(
+        data.title,
+        maxLines: 2,
+        textAlign: TextAlign.center,
+        overflow: TextOverflow.ellipsis,
+        style: GoogleFonts.poppins(
+          color: red002,
+          fontSize: data.title.length > 12 ? 17 : 25,
+          fontWeight: FontWeight.w700,
+        ),
+      ),
+      smokeyWhite,
+    );
+  }
+
+  duration() {
+    return TextContainer(
+      Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          SizedBox(width: 10),
+          Text(getDuration(), style: bodyMedium),
+          Icon(Icons.access_time, size: 15, color: Colors.white),
+        ],
+      ),
+      red002,
+    );
+  }
+
+  genre() {
+    return TextContainer(
+      Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          SizedBox(width: 10),
+          Text(capitalize(data.genre), style: bodyMedium),
+          Icon(Icons.movie, size: 15, color: Colors.white),
+        ],
+      ),
+      red002,
+    );
+  }
+
+  director() {
+    return TextContainer(
+      Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          SizedBox(width: 10),
+          Text(data.director, style: bodyMedium),
+          Icon(Icons.person, size: 15, color: Colors.white),
+        ],
+      ),
+      red002,
+    );
+  }
+
+  button() {
+    return Container(
+      decoration: BoxDecoration(
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withAlpha(30),
+            blurRadius: 5,
+            offset: const Offset(0, 3),
+          ),
+        ],
+      ),
+      child: FilledButton(
+        onPressed: () {},
+        style: FilledButton.styleFrom(
+          padding: const EdgeInsets.all(0),
+          backgroundColor: red002,
+          fixedSize: Size(108, 10),
+        ),
+        child:
+            Text(time != null ? "Nopirkt biļeti" : "Vairāk", style: bodyLarge),
+      ),
+    );
+  }
+
+  TextContainer(text, Color color) {
+    return Container(
+      width: 206,
+      margin: const EdgeInsets.symmetric(vertical: 5),
+      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 5),
+      decoration: BoxDecoration(
+        color: color,
+        borderRadius: BorderRadius.only(
+          topRight: Radius.circular(5),
+          bottomRight: Radius.circular(5),
+        ),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withAlpha(60),
+            blurRadius: 8,
+            offset: const Offset(0, 7),
+          ),
+        ],
+      ),
+      child: text,
+    );
+  }
+}
